Ohad's part
read/write to files.
the files are included movies.dat and sessions.dat
all five functions in interface are implmented for both entities
exceptions are added
injection from xml file and annoniations
postconstruct and predestroyer

all that is left is part 4
and the whole project needs to be tested because there will probally be some bugs.